requirements.txt
eth-account==0.13.7
python-dotenv==1.0.1
cerebras-cloud-sdk
fastapi
uvicorn
